refactor: move code from IndexPage to Video
