+++ frontend/src/components/Video/Video.tsx
@@ -1,22 +1,46 @@
+import { useCallback, useEffect, useState } from "react";
 import styles from "./Video.module.css";
 
-interface VideoProps {
+interface VideoInterface {
+  id: number;
+  title: string;
   url: string;
+  thumbnailUrl: string;
+  publish_date: string;
+  todays: boolean;
 }
 
-const Video = (props: VideoProps) => {
+const Video = () => {
+  const [video, setVideo] = useState<VideoInterface>();
+
+  const getVideo = useCallback(async () => {
+    const response = await fetch(
+      "https://wersow-api.herokuapp.com/api/videos/todays"
+    );
+    const data = await response.json();
+    setVideo(await data);
+  }, []);
+
+  useEffect(() => {
+    getVideo();
+  }, [getVideo]);
+
   return (
     <div className={styles.videoWrapper}>
-      <iframe
-        className={styles.video}
-        width="560"
-        height="315"
-        src={props.url.replace("watch?v=", "embed/")}
-        title="YouTube video player"
-        frameBorder="0"
-        allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-        allowFullScreen
-      ></iframe>
+      {video ? (
+        <iframe
+          className={styles.video}
+          width="560"
+          height="315"
+          src={video.url.replace("watch?v=", "embed/")}
+          title="YouTube video player"
+          frameBorder="0"
+          allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
+          allowFullScreen
+        ></iframe>
+      ) : (
+        <></>
+      )}
     </div>
   );
 };
